# api/openapi.yaml
openapi: 3.0.1
info:
  title: VMClarity APIs
  version: 1.0.0
paths:
...
  /assetScans:
    get:
      summary: Get asset scans according to the given filters
      operationId: GetAssetScans
      parameters:
        - $ref: '#/components/parameters/odataFilter'
        - $ref: '#/components/parameters/odataSelect'
        - $ref: '#/components/parameters/odataCount'
        - $ref: '#/components/parameters/odataTop'
        - $ref: '#/components/parameters/odataSkip'
        - $ref: '#/components/parameters/odataExpand'
        - $ref: '#/components/parameters/odataOrderBy'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScans'

        default:
          $ref: '#/components/responses/UnknownError'
    post:
      summary: Create an asset scan for a specified asset
      operationId: PostAssetScans
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetScan'
        required: true
      responses:
        201:
          description: AssetScan was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScan'
        400:
          description: Invalid asset scan supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        409:
          description: Conflicting asset scan found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanExists'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
  /assetScans/{assetScanID}:
    get:
      summary: Get an asset scan.
      operationId: GetAssetScansAssetScanID
      parameters:
        - $ref: '#/components/parameters/assetScanID'
        - $ref: '#/components/parameters/odataSelect'
        - $ref: '#/components/parameters/odataExpand'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScan'
        404:
          description: AssetScan ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
    put:
      summary: Update an asset scan.
      operationId: PutAssetScansAssetScanID
      parameters:
        - $ref: '#/components/parameters/assetScanID'
        - $ref: '#/components/parameters/ifmatch'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetScan'
        required: true
      responses:
        200:
          description: Updated an asset scan successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScan'
        400:
          description: Invalid asset scan supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        404:
          description: AssetScan ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        409:
          description: Conflicting asset scan found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanExists'
        412:
          description: Etag didn't match.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
    patch:
      summary: Patch an asset scan
      operationId: PatchAssetScansAssetScanID
      parameters:
        - $ref: '#/components/parameters/assetScanID'
        - $ref: '#/components/parameters/ifmatch'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetScan'
        required: true
      responses:
        200:
          description: Updated an asset scan successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScan'
        400:
          description: Invalid asset scan supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        404:
          description: AssetScan ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        409:
          description: Conflicting asset scan found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetScanExists'
        412:
          description: Etag didn't match.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        message:
          type: string
          readOnly: true
      description: An object that is returned in all cases of failures.
...
    Metadata:
      type: object
      properties:
        annotations:
          $ref: '#/components/schemas/Annotations'
    MetadataReadOnly:
      type: object
      properties:
        annotations:
          $ref: '#/components/schemas/Annotations'
          readOnly: true
    Annotations:
      type: array
      description: Generic map of string keys and string values to attach arbitrary non-identifying metadata to objects.
      items:
        type: object
        properties:
          key:
            type: string
          value:
            type: string
...
    ScanRelationship:
      type: object
      description: Describes an expandable relationship to Scan object
      allOf:
        - $ref: '#/components/schemas/MetadataReadOnly'
        - type: object
          properties:
            id:
              type: string
            name:
              type: string
              readOnly: true
            revision:
              type: integer
              readOnly: true
            startTime:
              type: string
              format: date-time
              readOnly: true
            endTime:
              type: string
              format: date-time
              readOnly: true
            scanConfig:
              $ref: '#/components/schemas/ScanConfigRelationship'
              readOnly: true
            scope:
              description: |
                The query used to limit the scope of this scan. It uses
                the ODATA $filter query language to limit the collection of assets
                that this scan will operate over. For example
                `startswith(assetInfo.location, 'eu-west-2')` will limit this scan
                to just assets in the eu-west-2 AWS region.
              type: string
            assetScanTemplate:
              $ref: '#/components/schemas/AssetScanTemplate'
              readOnly: true
            assetIDs:
              description: List of asset IDs to be scanned
              type: array
              items:
                type: string
              nullable: true
              readOnly: true
            status:
              $ref: '#/components/schemas/ScanStatus'
              readOnly: true
            summary:
              $ref: '#/components/schemas/ScanSummary'
              readOnly: true
          required:
            - id
        - $ref: '#/components/schemas/ScanTemplateReadOnly'
...
    ScanFindingsSummary:
      description: A summary of the scan findings.
      type: object
      properties:
        totalPackages:
          type: integer
        totalExploits:
          type: integer
        totalMalware:
          type: integer
        totalMisconfigurations:
          type: integer
        totalRootkits:
          type: integer
        totalSecrets:
          type: integer
        totalInfoFinder:
          type: integer
        totalVulnerabilities:
          $ref: '#/components/schemas/VulnerabilityScanSummary'
...
    AssetRelationship:
      type: object
      description: Describes a relationship to an asset which can be expanded.
      allOf:
        - $ref: '#/components/schemas/MetadataReadOnly'
        - type: object
          properties:
            id:
              type: string
            revision:
              type: integer
              readOnly: true
            assetInfo:
              $ref: '#/components/schemas/AssetType'
              readOnly: true
            scansCount:
              description: Total number of scans that have ever run for this asset
              type: integer
              readOnly: true
            firstSeen:
              type: string
              format: date-time
            terminatedOn:
              type: string
              format: date-time
              readOnly: true
            lastSeen:
              type: string
              format: date-time
              readOnly: true
            summary:
              $ref: '#/components/schemas/ScanFindingsSummary'
              readOnly: true
            providers:
              type: array
              description: List of providers which discover the asset.
              items:
                $ref: '#/components/schemas/ProviderRelationship'
              readOnly: true
          required: ['id']
...
    AssetScans:
      type: object
      properties:
        count:
          type: integer
          description: Total asset scans count according to the given filters
          readOnly: true
        items:
          type: array
          description: List of asset scans according to the given filters and page.
            List length must be lower or equal to pageSize.
          items:
            $ref: '#/components/schemas/AssetScan'
          readOnly: true
    ResourceCleanupStatus:
      type: object
      required:
        - state
        - reason
        - lastTransitionTime
      properties:
        state:
          description: |
            Describes the state of resource cleanup.
            
            | State   | Description                                                |
            | ------- | ---------------------------------------------------------- |
            | Pending | Initial state for cleaning up resources                    |
            | Skipped | Resource cleanup has been skipped due to Delete Job Policy |
            | Failed  | Cleaning up resources has been failed                      |
            | Done    | Resources have been successfully cleaned up                |
          type: string
          enum:
            - Pending
            - Skipped
            - Failed
            - Done
        reason:
          description: |
            Machine readable reason for state transition.

            | State   | Reason          | Description                                                            |
            | ------- | --------------- | ---------------------------------------------------------------------- |
            | Pending | AssetScanCreate | AssetScan created                                                      |
            | Skipped | NotApplicable   | Resource cleanup is not required as it is managed outside of VMClarity |
            | Skipped | DeletePolicy    | Resource cleanup has been skipped due to Delete Job Policy             |
            | Failed  | ProviderError   | Failed due to Provider error                                           |
            | Failed  | InternalError   | Failed due to internal error                                           |
            | Done    | Success         | Successfully completed                                                 |
          type: string
          enum:
            - AssetScanCreated
            - NotApplicable
            - DeletePolicy
            - ProviderError
            - InternalError
            - Success
        message:
          description: Human readable message.
          type: string
        lastTransitionTime:
          description: Last date time when the status has changed.
          type: string
          format: date-time
    AssetScanTemplate:
      type: object
      properties:
        scanFamiliesConfig:
          $ref: '#/components/schemas/ScanFamiliesConfig'
        scannerInstanceCreationConfig:
          $ref: '#/components/schemas/ScannerInstanceCreationConfig'
    AssetScan:
      type: object
      allOf:
        - $ref: '#/components/schemas/Metadata'
        - type: object
          properties:
            id:
              type: string
              # TODO(sambetts) Decide if we want the validation here by having
              # separate schemas for GET, POST and PATCH.
              #
              # We can't make anything readOnly or required here if we want to
              # use the same schema for PATCH otherwise it will require you
              # always send the required fields and remove the readOnly fields
              # on the client side.
              #
              # readOnly: true
            revision:
              type: integer
            asset:
              $ref: '#/components/schemas/AssetRelationship'
            scan:
              $ref: '#/components/schemas/ScanRelationship'
            status:
              $ref: '#/components/schemas/AssetScanStatus'
            sbom:
              $ref: '#/components/schemas/SbomScan'
            vulnerabilities:
              $ref: '#/components/schemas/VulnerabilityScan'
            malware:
              $ref: '#/components/schemas/MalwareScan'
            rootkits:
              $ref: '#/components/schemas/RootkitScan'
            secrets:
              $ref: '#/components/schemas/SecretScan'
            misconfigurations:
              $ref: '#/components/schemas/MisconfigurationScan'
            exploits:
              $ref: '#/components/schemas/ExploitScan'
            infoFinder:
              $ref: '#/components/schemas/InfoFinderScan'
            findingsProcessed:
              type: boolean
            resourceCleanupStatus:
              $ref: '#/components/schemas/ResourceCleanupStatus'
            stats:
              $ref: '#/components/schemas/AssetScanStats'
            summary:
              $ref: '#/components/schemas/ScanFindingsSummary'
            provider:
              $ref: "#/components/schemas/ProviderRelationship"
              description: Provider responsible for picking up and running the asset scan.
        - $ref: '#/components/schemas/AssetScanTemplate'
      # TODO(sambetts) Decide if we want the validation here by having
      # separate schemas for GET, POST and PATCH.
      #
      # We can't make anything readOnly or required here if we want to
      # use the same schema for PATCH otherwise it will require you
      # always send the required fields and remove the readOnly fields
      # on the client side.
      #
      # required:
      #  - asset
      #  - scan
    AssetScanStatus:
      type: object
      required:
        - state
        - reason
        - lastTransitionTime
      properties:
        state:
          description: |
            Describes the state of scan result.

            | State       | Description                                                                                      |
            | ----------- | ------------------------------------------------------------------------------------------------ |
            | Pending     | Initial state for AssetScan waiting for being scheduled                                          |
            | Scheduled   | AssetScan which has been scheduled on Provider                                                   |
            | ReadyToScan | Provider acknowledged that scanners for AssetScan is ready to run                                |
            | InProgress  | Scanners are running on the Target                                                               |
            | Aborted     | AssetScan has been aborted and all running Scanners need to be cancelled and shutdown gracefully |
            | Failed      | Running Scanners on Target has failed, check *reason* and *message* fields for the details       |
            | Done        | Running Scanners on Target has finished with no errors                                           |
          type: string
          enum:
            - Pending
            - Scheduled
            - ReadyToScan
            - InProgress
            - Aborted
            - Failed
            - Done
        reason:
          description: |
            Machine readable reason for state transition.

            | State       | Reason           | Description                                                                                |
            | ----------- | ---------------- | ------------------------------------------------------------------------------------------ |
            | Pending     | Created          | Scan is pending as it has been newly created                                               |
            | Scheduled   | Provisioning     | Scan is scheduled as Asset resources are being provisioned                                 |
            | ReadyToScan | UnSupervised     | Scan is created manually without the involvement of the Orchestrator (e.g. CI/CD scenario) |
            | ReadyToScan | ResourcesReady   | Scan is ready to scan as Asset resources have been provisioned                             |
            | InProgress  | ScannerIsRunning | Scan is in progress as Scanner is still performing scanning on Asset resources             |
            | Aborted     | Cancellation     | Scan has been aborted due to cancellation request                                          |
            | Aborted     | Timeout          | Scan has been aborted due to timeout                                                       |
            | Failed      | Error            | Scan has failed to one or more errors reported by the Scanner                              |
            | Failed      | AbortTimeout     | Scan has failed due to being in Aborted state for too long                                 |
            | Done        | Success          | Scan has finished successfully                                                             |
          type: string
          enum:
            - Created
            - Provisioning
            - UnSupervised
            - ResourcesReady
            - ScannerIsRunning
            - Cancellation
            - Timeout
            - Error
            - AbortTimedOut
            - Success
          x-enum-varnames:
            - AssetScanStatusReasonCreated
            - AssetScanStatusReasonProvisioning
            - AssetScanStatusReasonUnSupervised
            - AssetScanStatusReasonResourcesReady
            - AssetScanStatusReasonScannerIsRunning
            - AssetScanStatusReasonCancellation
            - AssetScanStatusReasonTimeout
            - AssetScanStatusReasonError
            - AssetScanStatusReasonAbortTimeout
            - AssetScanStatusReasonSuccess
        message:
          description: Human readable message.
          type: string
        lastTransitionTime:
          description: Last date time when the status has changed.
          type: string
          format: date-time
...
    AssetScanStats:
      type: object
      properties:
        general:
          $ref: '#/components/schemas/AssetScanGeneralStats'
        sbom:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        malware:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        rootkits:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        secrets:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        misconfigurations:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        exploits:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        infoFinder:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
...
    AssetScanExists:
      type: object
      properties:
        message:
          description: Describes which unique constraint combination causes the conflict.
          readOnly: true
          type: string
        assetScan:
          $ref: '#/components/schemas/AssetScan'
...
    ProviderRelationship:
      type: object
      description: Describes a relationship to a provider which can be expanded.
      allOf:
        - $ref: '#/components/schemas/MetadataReadOnly'
        - type: object
          properties:
            id:
              type: string
            revision:
              type: integer
              readOnly: true
            displayName:
              description: Human-readable name for the provider.
              type: string
              readOnly: true
            lastHeartbeatTime:
              description: Timestamp of the last heartbeat from the provider.
              type: string
              format: date-time
              readOnly: true
            status:
              $ref: '#/components/schemas/ProviderStatus'
              readOnly: true
            providerRuntimeVersion:
              description: Version of the provider runtime that the provider was compiled with.
              type: string
              readOnly: true
...
    SbomScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        packages:
          type: array
          items:
            $ref: '#/components/schemas/Package'
          nullable: true
    VulnerabilityScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/Vulnerability'
          nullable: true
    MalwareScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        malware:
          type: array
          items:
            $ref: '#/components/schemas/Malware'
          nullable: true
        metadata:
          type: array
          items:
            $ref: '#/components/schemas/ScannerMetadata'
          nullable: true
    RootkitScan:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ScannerStatus'
        rootkits:
          type: array
          items:
            $ref: '#/components/schemas/Rootkit'
          nullable: true
    SecretScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        secrets:
          type: array
          items:
            $ref: '#/components/schemas/Secret'
          nullable: true
    MisconfigurationScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        scanners:
          type: array
          items:
            type: string
          nullable: true
        misconfigurations:
          type: array
          items:
            $ref: '#/components/schemas/Misconfiguration'
          nullable: true
    InfoFinderScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        scanners:
          type: array
          items:
            type: string
          nullable: true
        infos:
          type: array
          items:
            $ref: '#/components/schemas/InfoFinderInfo'
          nullable: true
...
    ExploitScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        exploits:
          type: array
          items:
            $ref: '#/components/schemas/Exploit'
          nullable: true
...
  responses:
...
    UnknownError:
      description: Unknown error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  parameters:
...
    odataFilter:
      name: "$filter"
      in: query
      schema:
        type: string
    odataSelect:
      name: "$select"
      in: query
      schema:
        type: string
    odataCount:
      name: "$count"
      in: query
      schema:
        type: boolean
    odataTop:
      name: "$top"
      in: query
      schema:
        type: integer
    odataSkip:
      name: "$skip"
      in: query
      schema:
        type: integer
    odataExpand:
      name: "$expand"
      in: query
      schema:
        type: string
    odataOrderBy:
      name: "$orderby"
      in: query
      schema:
        type: string
      x-go-name: "OrderBy"
...
    assetScanID:
      name: assetScanID
      in: path
      required: true
      schema:
        type: string
...
    ifmatch:
      name: If-Match
      in: header
      schema:
        type: integer
...